backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg==3.2.3
asyncpg==0.29.0
greenlet==3.0.3
pydantic==2.10.3
pydantic-settings==2.6.1
alembic==1.14.0
redis==5.2.0
pgvector==0.3.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
